Changed chkfilelen to use long long.
git-svn-id: file://localhost/data1/git/presto_svn/trunk@379 5fac236c-a0d8-0310-9745-ef116914eda1

+++ include/chkio.h
@@ -26,7 +26,7 @@ int chkfseek(FILE * stream, long offset, int whence);
 int chkfileseek(FILE * stream, long offset, size_t size, int whence);
 /* Preform a file seek with error checking.  */
 
-unsigned long chkfilelen(FILE *file, size_t size);
+long long chkfilelen(FILE *file, size_t size);
 /* Return the length of a file (in blocks of 'size').  */
 
 int read_int(FILE *infile);
